+++ src/helpers/fetchHeaders.ts
@@ -1,0 +1,3 @@
+export const fetchHeaders: Headers = new Headers({
+	'Api-User-Agent': 'Email: tientq64@gmail.com; GitHub: tientq64/taxon4'
+})