+++ src/helpers/makeAborter.ts
@@ -1,0 +1,6 @@
+export function makeAborter(): AbortController {
+	const aborter: AbortController = new AbortController()
+	aborter.abort = aborter.abort.bind(aborter)
+
+	return aborter
+}