+++ web-extension/helpers/switchToPage.ts
@@ -13,6 +13,14 @@ export function getCurrentSearchQuery(): string | undefined {
 		case sites.flickrSearch:
 			q = document.querySelector<HTMLInputElement>('#search-field')!.value
 			break
+
+		case sites.inaturalistSearch:
+			q = document.querySelector<HTMLInputElement>('#q')!.value
+			break
+
+		case sites.inaturalistTaxon:
+			q = document.querySelector<HTMLElement>('#TaxonHeader .sciname')!.innerText
+			break
 	}
 	if (q === undefined) return
 
@@ -20,11 +28,15 @@ export function getCurrentSearchQuery(): string | undefined {
 	return q
 }
 
-export function getSwitchToPageUrl(pageName: keyof Sites): string | undefined {
+export function switchToPage(pageName: keyof Sites): void
+export function switchToPage(pageName: 'inaturalistTaxon', isCommonName?: boolean): void
+
+export function switchToPage(pageName: keyof Sites, ...args: unknown[]): void {
+	let url: string | undefined = undefined
+
 	let q: string | undefined = getCurrentSearchQuery()
 	if (q === undefined) return
 
-	let url: string | undefined = undefined
 	switch (pageName) {
 		case 'wikipedia':
 			url = `https://en.wikipedia.org/wiki/${q}`
@@ -37,14 +49,14 @@ export function getSwitchToPageUrl(pageName: keyof Sites): string | undefined {
 		case 'inaturalistSearch':
 			url = `https://www.inaturalist.org/taxa/search?view=list&q=${q}`
 			break
+
+		case 'inaturalistTaxon':
+			{
+				const isCommonNameQuery: string = args[0] ? '&isCommonName' : ''
+				url = `https://www.inaturalist.org/taxa/search?view=list&q=${q}${isCommonNameQuery}`
+			}
+			break
 	}
-	if (url === undefined) return
-
-	return url
-}
-
-export function switchToPage(pageName: keyof Sites): void {
-	const url: string | undefined = getSwitchToPageUrl(pageName)
 	if (url === undefined) return
 
 	location.href = url