+++ web-extension/meta.user.js
@@ -4,11 +4,12 @@
 // @version      0.1.0
 // @author       tientq64
 // @description  Thu thập dữ liệu các loài sinh vật cho taxon4.
-// @icon         https://cdn-icons-png.flaticon.com/128/714/714000.png
+// @icon         https://cdn-icons-png.flaticon.com/128/5330/5330052.png
 // @match        *://*/*
 // @exclude      *://localhost:*/*
-// @exclude      *://www.youtube.com/*
-// @exclude      *://tailwindcss.com/*
+// @exclude      https://www.youtube.com/*
+// @exclude      https://tailwindcss.com/*
+// @exclude      https://chatgpt.com/*
 // @require      {scriptURL}
 // @resource     css {styleURL}
 // @grant        GM_getResourceText
@@ -18,5 +19,6 @@
 // @grant        GM_deleteValue
 // @grant        GM_openInTab
 // @grant        window.close
+// @grant        window.onurlchange
 // @noframes
 // ==/UserScript==